Abstracting away different parts of WhatsAppEditor
