+++ src/common/RichEditor.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import reactStringReplace from 'react-string-replace';
+import { draftToMarkdown } from 'markdown-draft-js';
+import { convertToRaw } from 'draft-js';
 
 // Indicates how to replace different parts of the text from WhatsApp to HTML.
 export const TextReplacements: any = [
@@ -44,8 +46,15 @@ export const TextReplacements: any = [
 // Finds double asterisks in text with a regular expression.
 export const findDoubleAsterisks = new RegExp(/\*{2}(.+?)\*{2}/g);
 
+// Convert Draft.js to WhatsApp message format.
+export const convertToWhatsApp = (editorState: any) => {
+  let markdownString = draftToMarkdown(convertToRaw(editorState.getCurrentContent()));
+  let messageText = markdownString.replace(findDoubleAsterisks, '*$1*');
+  return messageText;
+};
+
 // Converts WhatsApp message formatting into HTML elements.
-export const textConverter = (text: any) => {
+export const WhatsAppToJsx = (text: any) => {
   let replacements = TextReplacements;
   for (let i = 0; i < replacements.length; i++) {
     let type = Object.keys(replacements[i])[0];
